// src/app/model/Employee.model.ts
export class UserModel {
  // filter(arg0: (res: any) => any): UserModel {
  //   throw new Error('Method not implemented.');
  // }
  // id: number = 0;
  fullName: string = '';
  lastName: string = '';
  email: string = '';
  phone: string = '';
  userName: string = '';
  password: string = '';
  // creationDate: string = '';
}